api/requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
Django==3.1.3
django-apscheduler==0.5.1
django-cors-headers==3.5.0
idna==2.10
mysqlclient==2.0.1
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.2
